+++ apps/lundin/src/app/modules/ancestry/ancestry-list/ancestry-list.component.ts
@@ -1,8 +1,8 @@
 import { Component } from "@angular/core"
 import { Person } from "@lundin/api-interfaces"
-import { NavigationService } from "../../services/navigation.service"
-import { AddPersonComponent } from "./add-person/add-person.component"
-import { AncestryService } from "./ancestry.service"
+import { NavigationService } from "../../../services/navigation.service"
+import { AddPersonComponent } from "../add-person/add-person.component"
+import { AncestryService } from "../ancestry.service"
 
 @Component({
 	selector: "lundin-ancestry-list",
