+++ apps/lundin/src/app/modules/ancestry/ancestry.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from "@angular/core"
 import { BrowserModule } from "@angular/platform-browser"
 import { SharedModule } from "../../shared/shared.module"
 import { AddPersonComponent } from "./add-person/add-person.component"
-import { AncestryListComponent } from "./ancestry-list.component"
+import { AncestryListComponent } from "./ancestry-list/ancestry-list.component"
 import { AncestryService } from "./ancestry.service"
 
 @NgModule({
